+++ src/com/github/unchama/gui/GuiMenu.java
@@ -6,12 +6,8 @@ import java.util.LinkedHashMap;
 import com.github.unchama.gui.minestack.MineStackMenuManager;
 import com.github.unchama.gui.moduler.GuiMenuManager;
 import com.github.unchama.gui.skillmenu.SkillTypeMenuManager;
-import com.github.unchama.gui.skillmenu.condendation.CondensationMenuManager;
 import com.github.unchama.gui.skillmenu.explosion.E_RangeMenuManager;
 import com.github.unchama.gui.skillmenu.explosion.ExplosionMenuManager;
-import com.github.unchama.gui.skillmenu.fairyaegis.FairyAegisMenuManager;
-import com.github.unchama.gui.skillmenu.magicdrive.MagicDriveMenuManager;
-import com.github.unchama.gui.skillmenu.ruinfield.RuinFieldMenuManager;
 
 
 public final class GuiMenu {
@@ -20,10 +16,6 @@ public final class GuiMenu {
 		SKILLTYPEMENU(SkillTypeMenuManager.class),
 		EXPLOSIONMENU(ExplosionMenuManager.class),
 		E_RANGEMENU(E_RangeMenuManager.class),
-		MAGICDRIVEMENU(MagicDriveMenuManager.class),
-		CONDENSATIONMENU(CondensationMenuManager.class),
-		RUINFIELDMENU(RuinFieldMenuManager.class),
-		FAIRYAEGISMENU(FairyAegisMenuManager.class),
 		MINESTACKMENU(MineStackMenuManager.class),
 		;
